Remove unused oldui
