--- app/ui/oldui.py
@@ -1,28 +1,0 @@
-from flask import Flask, render_template, request, jsonify
-from flask_cors import CORS
-import requests
-import json
-
-app = Flask(__name__)
-CORS(app)
-
-def send_message(message):
-    url = "http://localhost:8086/api/v1/user-message"
-    headers = {"Content-Type": "application/json"}
-    data = {"message": message}
-    response = requests.post(url, headers=headers, data=json.dumps(data))
-    if response.status_code == 200:
-        return response.text
-    else:
-        return f"Error: {response.status_code} - {response.text}"
-
-@app.route('/', methods=['GET', 'POST'])
-def index():
-    if request.method == 'POST':
-        user_message = request.form['message']
-        ai_response = send_message(user_message)
-        return render_template('index.html', ai_response=ai_response)
-    return render_template('index.html')
-
-if __name__ == '__main__':
-    app.run(debug=True, host='0.0.0.0', port=5000)
